// backend/src/types/auth.types.ts
export interface SignupRequest {
  name: string;
  email: string;
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt?: Date;
}

export interface AuthResponse {
  user: User;
  token: string;
}